Refactor auxiliaries
